requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.0.1